requirements.txt

notion-client  
pinecone  
flask  
python-dotenv  